+++ QuiperTests/AppTests.swift
@@ -46,6 +46,7 @@ class MockMainWindowController: MainWindowControlling {
     var window: NSWindow? = NSWindow()
     var currentWebViewURLToReturn: URL? = URL(string: "https://example.com")
     var activeServiceURL: String? = "https://example.com"
+    var activeServiceID: UUID?
     var activeWebView: WKWebView? = nil
     var focusInputInActiveWebviewCalled = false
     var focusInputInActiveWebviewWithFallbackCalled = false